+++ client/client.rb
@@ -10,7 +10,8 @@ require 'json'
 
 def create_header(body)
   key = "example"
-  signature = URI.encode_www_form(body)
+  #signature = URI.encode_www_form(body)
+  signature = body
   hmac = HMAC::SHA1.new(key)
   hmac.update(signature)
   {"Authentication" =>"hmac " + CGI.escape(Base64.encode64("#{hmac.digest}"))}
@@ -19,22 +20,11 @@ end
 c = HTTPClient.new
 
 # See what we got in the database
-puts c.get("http://localhost:8081/test").body
-
-# Insert something
-body = {"name" => "unrar", "age" => 17, "password" => "can_you_not!"}
+body = "users" # That's what we want to sign with HMAC
 header = create_header(body)
-res = c.post("http://localhost:8081/test/", body=body, header=header)
-# Parse the JSON to get the ID
-jsres = JSON.parse(res.body)
-id = jsres["id"]
+# The body=body part is useless but it won't work otherwise
+res = c.get("http://localhost:8081/users", body=body, header=header)
 puts res.body
-puts "id: #{id}"
 
-# Change the password
-body = {"password" => "yes_i_can!"}
-header = create_header(body)
-res = c.put("http://localhost:8081/test/#{id}", body=body, header=header)
-puts res.body